Fix more type casts.

+++ tests/keyentriessingletontests.cpp
@@ -81,10 +81,10 @@ void KeyEntriesSingletonTests::e2eTests()
     // Verify that all of the other values in the key entry are what we expect.
     QCOMPARE(QString("Test Key"), foundKey->identifier());
     QCOMPARE(QString("Test Issuer"), foundKey->issuer());
-    QCOMPARE((unsigned int)KEYENTRY_KEYTYPE_HEX, foundKey->keyType());
-    QCOMPARE((unsigned int)KEYENTRY_OTPTYPE_HOTP, foundKey->otpType());
+    QCOMPARE(KEYENTRY_KEYTYPE_HEX, foundKey->keyType());
+    QCOMPARE(KEYENTRY_OTPTYPE_HOTP, foundKey->otpType());
     QCOMPARE((unsigned int)30, foundKey->timeStep());
-    QCOMPARE((unsigned int)KEYENTRY_ALG_SHA1, foundKey->algorithm());
+    QCOMPARE(KEYENTRY_ALG_SHA1, foundKey->algorithm());
     QVERIFY(foundKey->codeValid());
     QCOMPARE((unsigned int)0, foundKey->timeOffset());
     QVERIFY(!foundKey->currentCode().isEmpty());        // Should be empty.  Not calculated yet.
